Rename NSDateInterval
In the iOS 10 SDKs NSDateInterval is a class provided by the SDK so redeclaring it in a framework throws errors.

This simply renames that class/file to a three letter namespaced version of the same file. Additionally, it adds the legacy Swift setting so peopl can build these frameworks with Xcode 8.